Move all users to the UserRepository
