+++ src/test/java/com/ite/itea/persistence/UserRepositoryTest.java
@@ -15,22 +15,30 @@ class UserRepositoryTest {
 
     private File file;
 
+    private UserRepository userRepository;
+
     private final ByteArrayOutputStream outContent = new ByteArrayOutputStream();
 
     @BeforeEach
     public void setUp() {
         System.setOut(new PrintStream(outContent));
         file = new File(Objects.requireNonNull(getClass().getClassLoader().getResource("usersWithOrders/users.txt")).getFile());
+        userRepository = new UserRepository(file);
     }
 
     @Test
-    void shouldReturnCaptainHookWhenAUserWithTheNameHookIsSearched() {
-        var userManagement = new UserManagement(file);
+    void shouldReturnAllUsersWhenAllUsersAreRequested() {
+        var users = userRepository.getAllUsers();
 
-        var user = userManagement.getUserByLastname("Hook");
-
-        then(user.firstname()).isEqualTo("Captain");
-        then(user.lastname()).isEqualTo("Hook");
+        then(users.size()).isEqualTo(4);
+        then(users.get(0).firstname()).isEqualTo("Peter");
+        then(users.get(0).lastname()).isEqualTo("Pan");
+        then(users.get(1).firstname()).isEqualTo("Captain");
+        then(users.get(1).lastname()).isEqualTo("Hook");
+        then(users.get(2).firstname()).isEqualTo("Tinker");
+        then(users.get(2).lastname()).isEqualTo("Bell");
+        then(users.get(3).firstname()).isEqualTo("Lost");
+        then(users.get(3).lastname()).isEqualTo("Boys");
     }
 
 }
